feat: fix encryptor module
