--- main.py
@@ -1,27 +1,0 @@
-import os
-
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives import padding as c_padding
-
-
-def encrypt_data(data, password):
-    salt = os.urandom(16)
-    kdf = PBKDF2HMAC(
-        algorithm=hashes.SHA256(),
-        length=32,
-        salt=salt,
-        iterations=100000,
-        backend=default_backend()
-    )
-    key = kdf.derive(password.encode())
-
-    iv = os.urandom(16)
-    cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=default_backend())
-    encryptor = cipher.encryptor()
-    ctx = c_padding.PKCS7(128).padder().update(data) + c_padding.PKCS7(128).padder().finalize()
-    encrypted_data = encryptor.update(ctx) + encryptor.finalize()
-
-    return salt + iv + encrypted_data
